+++ mark1/src/main/java/com/starkindustries/dao/ProductMapper.java
@@ -3,13 +3,13 @@ package com.starkindustries.dao;
 import com.starkindustries.pojo.Product;
 
 public interface ProductMapper {
-    int deleteByPrimaryKey(Byte id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Product record);
 
     int insertSelective(Product record);
 
-    Product selectByPrimaryKey(Byte id);
+    Product selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(Product record);
