+++ mark1/src/main/java/com/starkindustries/dao/UserMapper.java
@@ -3,13 +3,13 @@ package com.starkindustries.dao;
 import com.starkindustries.pojo.User;
 
 public interface UserMapper {
-    int deleteByPrimaryKey(Byte id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(User record);
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Byte id);
+    User selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(User record);
