+++ mark1/src/main/java/com/starkindustries/pojo/User.java
@@ -3,7 +3,7 @@ package com.starkindustries.pojo;
 import java.util.Date;
 
 public class User {
-    private Byte id;
+    private Integer id;
 
     private String username;
 
@@ -17,13 +17,13 @@ public class User {
 
     private String answer;
 
-    private Byte role;
+    private Integer role;
 
     private Date createtime;
 
     private Date updatetime;
 
-    public User(Byte id, String username, String password, String email, String phone, String question, String answer, Byte role, Date createtime, Date updatetime) {
+    public User(Integer id, String username, String password, String email, String phone, String question, String answer, Integer role, Date createtime, Date updatetime) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -40,11 +40,11 @@ public class User {
         super();
     }
 
-    public Byte getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Byte id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -96,11 +96,11 @@ public class User {
         this.answer = answer == null ? null : answer.trim();
     }
 
-    public Byte getRole() {
+    public Integer getRole() {
         return role;
     }
 
-    public void setRole(Byte role) {
+    public void setRole(Integer role) {
         this.role = role;
     }
